SPA 2 Complete
Edit - Tested with different inputs and fixed logic

Issues found and corrected:
1. In-place editing of the dictionary of possible answers led to errors in if-logic down the line (pointers are tricky like that).
2. can_match never checked if a letter with a ? wasn't in the same position as the ?, allowing incorrect answers to slip through.

In the future:
We were required to use only basic arrays for this lab. There are many points in this lab where the implementation of a vector or hash set data structure would have proved to be more effective than an array both in terms of big O runtime and memory efficiency.